feat(preview-middleware): export initAdp function
