+++ packages/preview-middleware/src/ui5/middleware.ts
@@ -1,49 +1,8 @@
 import { LogLevel, ToolsLogger, UI5ToolingTransport } from '@sap-ux/logger';
 import type { RequestHandler } from 'express';
-import type { MiddlewareParameters, MiddlewareUtils } from '@ui5/server';
-import { FlpSandbox } from '../base/flp';
+import type { MiddlewareParameters } from '@ui5/server';
+import { FlpSandbox, initAdp } from '../base/flp';
 import type { MiddlewareConfig } from '../types';
-import { AdpPreview, type AdpPreviewConfig } from '@sap-ux/adp-tooling';
-import type { ReaderCollection } from '@ui5/fs';
-
-/**
- * Initialize the preview for an adaptation project.
- *
- * @param rootProject reference to the project
- * @param config configuration from the ui5.yaml
- * @param flp FlpSandbox instance
- * @param util middleware utilities provided by the UI5 CLI
- * @param logger logger instance
- */
-async function initAdp(
-    rootProject: ReaderCollection,
-    config: AdpPreviewConfig,
-    flp: FlpSandbox,
-    util: MiddlewareUtils,
-    logger: ToolsLogger
-) {
-    const appVariant = await rootProject.byPath('/manifest.appdescr_variant');
-    if (appVariant) {
-        const adp = new AdpPreview(config, rootProject, util, logger);
-        const variant = JSON.parse(await appVariant.getString());
-        const layer = await adp.init(variant);
-        if (flp.rta) {
-            flp.rta.layer = layer;
-            flp.rta.options = {
-                projectId: variant.id,
-                scenario: 'ADAPTATION_PROJECT'
-            };
-            for (const editor of flp.rta.editors) {
-                editor.pluginScript ??= 'open/ux/preview/client/adp/init';
-            }
-        }
-        await flp.init(adp.descriptor.manifest, adp.descriptor.name, adp.resources);
-        flp.router.use(adp.descriptor.url, adp.proxy.bind(adp) as RequestHandler);
-        adp.addApis(flp.router);
-    } else {
-        throw new Error('ADP configured but no manifest.appdescr_variant found.');
-    }
-}
 
 /**
  * The developer mode is only supported for adaptation projects, therefore, notify the user if it is wrongly configured and then disable it.
